Refactor: addEvent, deleteEvent 리팩토링
