+++ src/components/AddEventModal/AddEventModal.style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const EventModalWrapper = styled.div`
+export const AddEventModalWrapper = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -9,6 +9,13 @@ export const EventModalWrapper = styled.div`
   padding-top: 5%;
 `;
 
+// export const ModalHeader = styled.div`
+//   width: 100%;
+//   height: 10%;
+//   text-align: center;
+//   font-size: 25px;
+// `;
+
 export const EventContainer = styled.div`
   display: flex;
   flex-direction: column;
@@ -16,35 +23,21 @@ export const EventContainer = styled.div`
   height: fit-content;
   margin-bottom: 5%;
 `;
-export const EventTitle = styled.div`
+export const EventTitle = styled.input`
   width: 100%;
   height: 30px;
-  border: 1px solid black;
-  border-radius: 3px;
-  padding: 2px 3px;
-  color: black;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  word-break: break-all;
 
   &:focus {
     outline: none;
   }
 `;
 
-export const EventDescription = styled.div`
+export const EventDescription = styled.textarea`
   width: 100%;
   height: 110px;
-  padding: 4px;
-  border: 1px solid black;
-  border-radius: 3px;
-  color: black;
-  overflow: scroll;
-  &:active {
+
+  &:focus {
     outline: none;
-  }
-  &::-webkit-scrollbar {
-    display: none;
   }
 `;
 
@@ -62,34 +55,20 @@ export const EventDateWrapper = styled.div`
   width: 100%;
   height: 100%;
 `;
-export const InputLabel = styled.p`
-  display: inline-block;
-  margin-bottom: 0.5rem;
-`;
-export const EventDate = styled.div`
+export const InputLabel = styled.label``;
+export const EventDate = styled.input`
   width: 90%;
   height: fit-content;
   text-align: center;
-  border: 1px solid black;
-  border-radius: 3px;
-  padding: 2px;
-  color: black;
   &:focus {
     outline: none;
   }
 `;
-
-export const ButtonWrapper = styled.div`
-  display: flex;
-  width: 150px;
-  justify-content: space-around;
-`;
-
 export const Button = styled.button`
   display: flex;
   align-items: center;
   justify-content: center;
-  width: 60px;
+  width: 70px;
   height: 30px;
   margin: 0 auto;
   margin-top: 30px;
@@ -98,8 +77,13 @@ export const Button = styled.button`
   background-color: #f59c00;
   color: white;
   font-size: 16px;
+  outline: none;
   cursor: pointer;
   &:hover {
     filter: brightness(80%);
+  }
+
+  &:active {
+    outline: none;
   }
 `;
