+++ src/pages/MainPage.jsx
@@ -6,37 +6,15 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import Modal from 'react-modal';
 import mainpageAPI from '../api/mainpageAPI';
-import EventModal from '../components/EventModal';
 
 const MainPage = () => {
   const [teamInfo, setTeamInfo] = useState([]);
   const [teamMembers, setTeamMembers] = useState();
   const [isStudentListModalOpen, setIsStudentListModalOpen] = useState(false);
-  const [isAddEventModalOpen, setIsAddEventModalOpen] = useState(false);
-  const [isEventModalOpen, setIsEventModalOpen] = useState(false);
 
   const [studentList, setStudentList] = useState([]);
   const [checkedList, setCheckedList] = useState([]);
   const [isChecked, setIsChecked] = useState(false);
-
-  const [event, setEvent] = useState({
-    title: '',
-    start: '',
-    end: '',
-    description: '',
-    teamId: null,
-  });
-  const [eventList, setEventList] = useState([
-    {
-      id: '',
-      title: '',
-      description: '',
-      start: '',
-      end: '',
-    },
-  ]);
-
-  const [selectedEventId, setSelectedEventId] = useState('');
 
   const params = useParams();
   const teamId = Number(params.teamId.substring(1));
@@ -53,9 +31,9 @@ const MainPage = () => {
     refreshTeamInfo();
   };
 
-  const handleClickAddEvent = () => {
-    setIsAddEventModalOpen((prev) => !prev);
-  };
+  // const handleClickAddEvent = () => {
+  //   setIsAddEventModalOpen((prev) => !prev);
+  // };
 
   //+버튼 눌렀을 때 나오는 모달에서 사용하는 get메소드
   const showStudentList = () => {
@@ -103,18 +81,6 @@ const MainPage = () => {
     });
   };
 
-  const addEvent = () => {
-    mainpageAPI.addEvent(event).then((data) => {
-      console.log(data);
-    });
-  };
-
-  const getTeamEventList = () => {
-    mainpageAPI.getTeamEventList(teamId).then((data) => {
-      console.log(`getTeamEventList: ${data}`);
-      setEventList(data);
-    });
-  };
   //모달 style
   const StudentListModalStyle = {
     overlay: {
@@ -143,178 +109,14 @@ const MainPage = () => {
     },
   };
 
-  const AddEventModalStyle = {
-    overlay: {
-      backgroundColor: ' rgba(0, 0, 0, 0.4)',
-      width: '100%',
-      height: '100vh',
-      zIndex: '10',
-      position: 'fixed',
-      top: '0',
-      left: '0',
-      zIndex: '9999',
-    },
-    content: {
-      width: '360px',
-      height: '430px',
-      zIndex: '150',
-      position: 'absolute',
-      top: '50%',
-      left: '50%',
-      transform: 'translate(-50%, -50%)',
-      borderRadius: '10px',
-      boxShadow: '2px 2px 2px rgba(0, 0, 0, 0.25)',
-      backgroundColor: 'white',
-      overflow: 'auto',
-    },
-  };
-
-  const ShowEventModalStyle = {
-    overlay: {
-      backgroundColor: ' rgba(0, 0, 0, 0.4)',
-      width: '100%',
-      height: '100vh',
-      zIndex: '10',
-      position: 'fixed',
-      top: '0',
-      left: '0',
-      zIndex: '9999',
-    },
-    content: {
-      width: '360px',
-      height: '430px',
-      zIndex: '150',
-      position: 'absolute',
-      top: '50%',
-      left: '50%',
-      transform: 'translate(-50%, -50%)',
-      borderRadius: '10px',
-      boxShadow: '2px 2px 2px rgba(0, 0, 0, 0.25)',
-      backgroundColor: 'white',
-      overflow: 'auto',
-    },
-  };
-
-  const handleChange = (e) => {
-    console.log('지금 startDate', event.startDate);
-    console.log('지금 endDate', event.endDate);
-    const { name, value } = e.target;
-    setEvent({
-      ...event,
-      teamId: teamId,
-      [name]: value,
-    });
-    console.log(name, value);
-  };
-
-  //입력값 초기화
-  const resetEventValue = () => {
-    setEvent({
-      title: '',
-      start: '',
-      end: '',
-      description: '',
-      teamId: null,
-    });
-  };
-
-  const handleAddEvent = () => {
-    //빈값입력시 오류
-    if (event.title === '' || event.start === '' || event.end === '') {
-      alert('입력하지 않은 값이 있습니다.');
-      resetEventValue();
-
-      //끝나는 날짜가 시작하는 날짜보다 값이 크면 안됨
-    } else if (new Date(event.start) > new Date(event.end)) {
-      alert('시간을 확인해주세요.');
-      resetEventValue();
-    } else {
-      // 이벤트 추가
-      addEvent();
-      resetEventValue();
-      setIsAddEventModalOpen(false);
-      getTeamEventList();
-    }
-  };
-
-  const handleClickEvent = () => {
-    setIsEventModalOpen((prev) => !prev);
-  };
-
   useEffect(() => {
     refreshTeamInfo();
-  }, [event]);
+  }, []);
 
   return (
     <s.Wrapper>
       <s.LeftContainer>
-        <Calendar
-          // teamId={teamId}
-          setIsAddEventModalOpen={setIsAddEventModalOpen}
-          isAddEventModalOpen={isAddEventModalOpen}
-          // event={event}
-          // setEvent={setEvent}
-          // isEventModalOpen={isEventModalOpen}
-          // setIsEventModalOpen={setIsEventModalOpen}
-          handleClickEvent={handleClickEvent}
-          // selectedEventId={selectedEventId}
-          setSelectedEventId={setSelectedEventId}
-          getTeamEventList={getTeamEventList}
-          eventList={eventList}
-        />
-        <Modal
-          isOpen={isAddEventModalOpen}
-          style={AddEventModalStyle}
-          onRequestClose={handleClickAddEvent} // 오버레이나 esc를 누르면 핸들러 동작
-          ariaHideApp={false}
-        >
-          <s.AddEventModalWrapper>
-            <s.EventContainer>
-              <s.InputLabel htmlFor='title'>일정 제목</s.InputLabel>
-              <s.EventTitle type='text' name='title' onChange={handleChange} />
-            </s.EventContainer>
-            <s.EventDateContainer>
-              <s.EventDateWrapper>
-                <s.InputLabel
-                  htmlFor='startDate'
-                  onChange={handleChange}
-                  // defaultValue={
-                  //   event.startDate !== '' && event.startDate
-                  // new Date('2023-11-23')
-                  // }
-                  // key={event.startDate}
-                >
-                  시작 날짜
-                </s.InputLabel>
-                <s.EventDate type='date' name='start' onChange={handleChange} />
-              </s.EventDateWrapper>
-              <s.EventDateWrapper>
-                <s.InputLabel htmlFor='endDate'>종료 날짜</s.InputLabel>
-                <s.EventDate type='date' name='end' onChange={handleChange} />
-              </s.EventDateWrapper>
-            </s.EventDateContainer>
-            <s.EventContainer>
-              <s.InputLabel for='description'>일정 설명</s.InputLabel>
-              <s.EventDescription
-                type='text'
-                name='description'
-                onChange={handleChange}
-              />
-            </s.EventContainer>
-            <s.Button onClick={handleAddEvent}>추가</s.Button>
-          </s.AddEventModalWrapper>
-        </Modal>
-        <Modal
-          isOpen={isEventModalOpen}
-          style={ShowEventModalStyle}
-          onRequestClose={handleClickEvent} // 오버레이나 esc를 누르면 핸들러 동작
-          ariaHideApp={false}
-        >
-          <EventModal
-            selectedEventId={selectedEventId}
-            getTeamEventList={getTeamEventList}
-          />
-        </Modal>
+        <Calendar teamId={teamId} />
       </s.LeftContainer>
       <s.RightContainer>
         {/* 팀원구성 */}
